+++ qjazz-server/qjazz-rpc/python/qjazz_rpc/cli.py
@@ -98,12 +98,6 @@ def print_config(conf: Optional[Path], pretty: bool = False):
     type=FilePathType,
 )
 @click.option("--force", is_flag=True, help="Force installation")
-@click.option(
-    "--list",
-    "list_plugins",
-    is_flag=True,
-    help="List installed plugins without installing them",
-)
 def install_plugins(configpath: Optional[Path], force: bool, list_plugins: bool = False):
     """Install plugins"""
     conf = load_configuration(configpath)
@@ -111,10 +105,7 @@ def install_plugins(configpath: Optional[Path], force: bool, list_plugins: bool 
 
     from qjazz_core.qgis import qgis_plugins
 
-    if list_plugins:
-        click.echo("Installed plugins:\n")
-        qgis_plugins._run_plugin_manager(conf.worker.qgis.plugins, "list")
-    elif force or conf.worker.qgis.plugins.install_mode == "auto":
+    if force or conf.worker.qgis.plugins.install_mode == "auto":
         qgis_plugins.install_plugins(conf.worker.qgis.plugins)
     else:
         click.echo("Plugin installation set to manual: no plugins to install...")